timob-1343: fixes based on comments
