--- android/modules/map/src/java/ti/modules/titanium/map/MapPointType.java
@@ -1,23 +1,0 @@
-package ti.modules.titanium.map;
-
-public class MapPointType {
-
-	private double latitude;
-	private double longitude;
-	
-	public MapPointType(double latitude, double longitude) 
-	{
-		this.latitude = latitude;
-		this.longitude = longitude;
-	}
-	
-	public double getLatitude() 
-	{
-		return latitude;
-	}
-	
-	public double getLongitude() 
-	{
-		return longitude;
-	}
-}
