+++ android/modules/map/src/java/ti/modules/titanium/map/ViewProxy.java
@@ -1,6 +1,6 @@
 /**
  * Appcelerator Titanium Mobile
- * Copyright (c) 2009-2011 by Appcelerator, Inc. All Rights Reserved.
+ * Copyright (c) 2009-2012 by Appcelerator, Inc. All Rights Reserved.
  * Licensed under the terms of the Apache Public License
  * Please see the LICENSE included with this distribution for details.
  */
@@ -51,7 +51,7 @@ public class ViewProxy extends TiViewProxy
 
 	private TiMapView mapView;
 	private ArrayList<AnnotationProxy> annotations;
-	private ArrayList<MapRouteType> routes;
+	private ArrayList<MapRoute> routes;
 	private ArrayList<TiMapView.SelectedAnnotation> selectedAnnotations;
 	
 	public ViewProxy()
@@ -64,7 +64,7 @@ public class ViewProxy extends TiViewProxy
 		//tiContext.addOnLifecycleEventListener(this);
 
 		annotations = new ArrayList<AnnotationProxy>();
-		routes = new ArrayList<MapRouteType>();
+		routes = new ArrayList<MapRoute>();
 		selectedAnnotations = new ArrayList<TiMapView.SelectedAnnotation>();
 	}
 
@@ -193,16 +193,18 @@ public class ViewProxy extends TiViewProxy
 			Object routeArray = routeMap.get("points");
 			if (routeArray instanceof Object[]) {
 				Object[] routes = (Object[]) routeArray;
-				MapPointType[] pointsType = new MapPointType[routes.length];
+				MapPoint[] pointsType = new MapPoint[routes.length];
 				for (int i = 0; i < routes.length; i++) {
+					
 					if (routes[i] instanceof HashMap) {
 						HashMap tempRoute = (HashMap)routes[i];
-						MapPointType mp = new MapPointType(TiConvert.toDouble(tempRoute, "latitude"), TiConvert.toDouble(tempRoute, "longitude"));
+						MapPoint mp = new MapPoint(TiConvert.toDouble(tempRoute, "latitude"), TiConvert.toDouble(tempRoute, "longitude"));
 						pointsType[i] = mp;
 					}
-					
 				}
-				MapRouteType mr = new MapRouteType(pointsType, TiConvert.toColor(routeMap, "color"), TiConvert.toInt(routeMap, "width"), TiConvert.toString(routeMap, "name"));
+				
+				MapRoute mr = new MapRoute(pointsType, TiConvert.toColor(routeMap, "color"), TiConvert.toInt(routeMap, "width"), TiConvert.toString(routeMap, "name"));
+				
 				if (mapView == null) {
 					this.routes.add(mr);
 				} else {
